Added classes deleted by mistake, fixed names and added new packages.
