+++ src/main/java/org/example/HTTP/ClassResources.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.HTTP;
 
 import java.io.IOException;
 import java.net.URI;
@@ -7,17 +7,17 @@ import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 
-public class RaceInformation {
-    private final String address = "https://www.dnd5eapi.co/api/races/:index";
+public class ClassResources {
+    private final String address = "https://www.dnd5eapi.co/api/classes/:index/levels/1";
     private final HttpClient client = HttpClient.newHttpClient();
 
-    public HttpResponse<String> getRaceInformation(String index) throws RuntimeException {
+    public HttpResponse<String> getClassResources(String index) {
         HttpRequest request;
         try {
             request = HttpRequest.newBuilder(new URI(address.replace(":index", index))).GET().build();
             HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
             return response;
-        } catch (IOException | URISyntaxException | InterruptedException | NullPointerException e) {
+        } catch (URISyntaxException | InterruptedException | IOException | NullPointerException e) {
             throw new RuntimeException(e);
         }
     }
