+++ src/main/java/org/example/HTTP/Proficiencies.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.HTTP;
 
 import java.io.IOException;
 import java.net.URI;
@@ -7,17 +7,17 @@ import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 
-public class ClassSpells {
-    private final String address = "https://www.dnd5eapi.co/api/classes/:index/levels/1/spells";
+public class Proficiencies {
+    private final String address = "https://www.dnd5eapi.co/api/races/:index/proficiencies";
     private final HttpClient client = HttpClient.newHttpClient();
 
-    public HttpResponse<String> getClassSpells(String index) {
+    public HttpResponse<String> getProficiencies(String index) {
         HttpRequest request;
         try {
             request = HttpRequest.newBuilder(new URI(address.replace(":index", index))).GET().build();
-            HttpResponse<String>response = client.send(request,HttpResponse.BodyHandlers.ofString());
+            HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
             return response;
-        } catch (URISyntaxException | InterruptedException | IOException | NullPointerException e) {
+        } catch (IOException | URISyntaxException | InterruptedException | NullPointerException e) {
             throw new RuntimeException(e);
         }
     }
