+++ src/main/java/org/example/MAPPERS/AbilityBonusMAPPER.java
@@ -1,8 +1,9 @@
-package org.example;
+package org.example.MAPPERS;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.example.StandardClasses.AbilityBonus;
 
 public class AbilityBonusMAPPER {
     private static final ObjectMapper MAPPER = new ObjectMapper();
