+++ src/main/java/org/example/MAPPERS/FeatureMAPPER.java
@@ -1,18 +1,19 @@
-package org.example;
+package org.example.MAPPERS;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.example.StandardClasses.Feature;
 
-public class SpellsMAPPER {
+public class FeatureMAPPER {
     private static final ObjectMapper MAPPER = new ObjectMapper();
 
-    public static Spells convertJSON(String JSON) {
+    public static Feature convertJSON(String JSON) {
         try {
-            Spells spells = new Spells();
+            Feature feature = new Feature();
             JsonNode node = MAPPER.readTree(JSON);
-            spells.setName((node.get("name").asText()));
-            return spells;
+            feature.setName((node.get("name").asText()));
+            return feature;
         } catch (JsonProcessingException e) {
             System.err.println(e.getMessage());
         }
