+++ src/main/java/org/example/MAPPERS/ItemMAPPER.java
@@ -1,8 +1,9 @@
-package org.example;
+package org.example.MAPPERS;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.example.StandardClasses.Item;
 
 public class ItemMAPPER {
     private static final ObjectMapper MAPPER = new ObjectMapper();
