+++ src/main/java/org/example/MAPPERS/SpellsMAPPER.java
@@ -1,18 +1,19 @@
-package org.example;
+package org.example.MAPPERS;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.example.StandardClasses.Spells;
 
-public class FeatureMAPPER {
+public class SpellsMAPPER {
     private static final ObjectMapper MAPPER = new ObjectMapper();
 
-    public static Feature convertJSON(String JSON) {
+    public static Spells convertJSON(String JSON) {
         try {
-            Feature feature = new Feature();
+            Spells spells = new Spells();
             JsonNode node = MAPPER.readTree(JSON);
-            feature.setName((node.get("name").asText()));
-            return feature;
+            spells.setName((node.get("name").asText()));
+            return spells;
         } catch (JsonProcessingException e) {
             System.err.println(e.getMessage());
         }
