+++ src/main/java/org/example/Main.java
@@ -10,34 +10,34 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Please, give the name of your character");
-        String answer = scanner.nextLine();
+        String name = scanner.nextLine();
 
-        String answer1;
+        String race;
         boolean existingRace = false;
         do {
             System.out.println("Please, give valid race name of your character");
-            answer1 = scanner.nextLine();
+            race = scanner.nextLine();
 
-            String response = RESTCaller.exists("https://www.dnd5eapi.co/api/races/%s".formatted(answer1));
+            String response = RESTCaller.exists("https://www.dnd5eapi.co/api/races/%s".formatted(race));
             if (response != null && response.trim().equals("OK")) {
                 existingRace = true;
             }
         } while (!existingRace);
 
-        String answer2;
+        String classes;
         boolean existingClass = false;
         do {
             System.out.println("Please, give the class of your character");
-            answer2 = scanner.nextLine();
+            classes = scanner.nextLine();
 
-            String response = RESTCaller.exists("https://www.dnd5eapi.co/api/classes/%s".formatted(answer2));
+            String response = RESTCaller.exists("https://www.dnd5eapi.co/api/classes/%s".formatted(classes));
             if (response != null && response.trim().equals("OK")) {
                 existingClass = true;
             }
         } while (!existingClass);
 
         String responseSpells = RESTCaller.passContent("https://www.dnd5eapi.co/api/classes/%s/spellcasting"
-                .formatted(answer2));
+                .formatted(classes));
         boolean spellsExist = responseSpells != null;
         if (spellsExist) {
             // TODO DEBUG INFO
@@ -45,16 +45,16 @@ public class Main {
             return;
         }
 
-        String answer3 = "";
+        String spell = "";
         if (spellsExist) {
             System.out.println("Please, select spells or/and cantrips of your character");
-            answer3 = scanner.nextLine();
+            spell = scanner.nextLine();
         }
 
         System.out.println("Please write background for your character");
-        String answer4 = scanner.nextLine();
+        String background = scanner.nextLine();
 
-        File file = new File("src\\main\\java\\org\\example\\" + answer + ".txt");
+        File file = new File("src\\main\\java\\org\\example\\" + name + ".txt");
         try {
             boolean created = file.createNewFile();
         } catch (IOException e) {
@@ -64,14 +64,14 @@ public class Main {
         try (FileWriter writer = new FileWriter(file); BufferedWriter bufferedWriter = new BufferedWriter(writer)) {
             bufferedWriter.flush();
 
-            bufferedWriter.write("race: " + answer1);
+            bufferedWriter.write("race: " + race);
             bufferedWriter.write("\n");
-            bufferedWriter.write("class: " + answer2);
+            bufferedWriter.write("class: " + classes);
             bufferedWriter.write(("\n"));
             //TODO user is able to select spells/cantrips if available (can be done via a list and number of the spell);
-            bufferedWriter.write("spells or/and cantrips: " + answer3);
+            bufferedWriter.write("spells or/and cantrips: " + spell);
             bufferedWriter.write(("\n"));
-            bufferedWriter.write("background: " + answer4);
+            bufferedWriter.write("background: " + background);
             bufferedWriter.write(("\n"));
         } catch (IOException ioException) {
             ioException.printStackTrace();
