+++ src/main/java/org/example/StandardClasses/ARaceByIndex.java
@@ -1,7 +1,6 @@
-package org.example;
+package org.example.StandardClasses;
 
 import java.util.List;
-import java.util.Objects;
 
 public class ARaceByIndex {
