+++ src/main/java/org/example/StandardClasses/AbilityBonus.java
@@ -1,0 +1,31 @@
+package org.example.StandardClasses;
+
+public class AbilityBonus {
+
+    private String name;
+    private int value;
+
+    public String getName() {
+        return this.name;
+    }
+
+    public int getValue() {
+        return this.value;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setValue(int value) {
+        this.value = value;
+    }
+
+    @Override
+    public String toString() {
+        return "{" +
+                "\"name\": \"" + name + "\", " +
+                "\"value\": " + value +
+                '}';
+    }
+}
