+++ src/main/java/org/example/StandardClasses/Language.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.StandardClasses;
 
 public class Language {
     private String index;
