+++ src/test/java/TestAbilityBonus.java
@@ -1,5 +1,5 @@
-import org.example.AbilityBonus;
-import org.example.AbilityBonusMAPPER;
+import org.example.StandardClasses.AbilityBonus;
+import org.example.MAPPERS.AbilityBonusMAPPER;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
