+++ src/test/java/TestClassSpells.java
@@ -1,6 +1,4 @@
-import org.example.ClassFeatures;
-import org.example.ClassSpells;
-import org.example.Proficiencies;
+import org.example.HTTP.ClassSpells;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
