+++ src/test/java/TestRaceInformation.java
@@ -1,7 +1,6 @@
-import org.example.RaceInformation;
+import org.example.HTTP.RaceInformation;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.opentest4j.AssertionFailedError;
 
 import java.net.http.HttpResponse;
